The Dice Problem
The Dice problem(functional)
Time Limit: 2 sec
Memory Limit: 128000 kB
Problem Statement
You are given a cubic dice with 6 faces. All the individual faces have numbers printed on them. The numbers are in the range of 1 to 6, like any ordinary dice. You will be provided with a face of this cube, your task is to find the number on the opposite face of the cube.

Note:- The sum of numbers on all opposite faces of the die is constant
Input
User Task:
Since this will be a functional problem, you don't have to take input. You just have to complete the function DiceProblem() that takes the integer N(face) as parameter.

Constraints:
1 <= N <= 6
Output
Return the number on the opposite side.
Example
Sample Input:-
2

Sample Output:-
5

Sample Input:-
1

Sample Output:-
6